Add Unmark as an external tool
`unmark` is a tool to produce an user manual from the `README.md`.

The `patch.json.gz` file contains a list of patterns to match and
their replacements.
If `unmark` complains about being unable to match a pattern, please open
an issue with the warning and your `README.md` at
[upstream](https://github.com/DMSalesman/Unmark).