Split GM9 across two sections in memory.
This is needed to implement support for Lua. Lua is way to big to fit in
a single contigious section of memory without overwriting our copy of
boot9 and boot11, so some clever linker manipulation was used to place
things in the free space after it.

This uses the following memory map:

```
AHBWRAM: 0x08000000 - 0x08080000
    .vectors
    .text
    .ARM.exidx

AHBWRAM2: 0x080a0000 - 0x08100000
    .text (files in arm9/source/lua)
    .rodata
    .data
    .bss
```